feat(Styles): flatter Styles dir structure
